addr-book-proj/requirements.txt
Flask==3.1.1
psycopg2-binary==2.9.10
Flask-CORS==3.0.10
gunicorn==22.0.0
python-dotenv==1.1.1